+++ shell/app/modules/msp/monitor/trace-insight/pages/trace-querier/trace-graph/span-title-info.tsx
@@ -1,0 +1,25 @@
+// Copyright (c) 2021 Terminus, Inc.
+//
+// This program is free software: you can use, redistribute, and/or modify
+// it under the terms of the GNU Affero General Public License, version 3
+// or later ("AGPL"), as published by the Free Software Foundation.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+// FITNESS FOR A PARTICULAR PURPOSE.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+import React from 'react';
+
+export function SpanTitleInfo(props: any) {
+  const { operationName, spanKind, component } = props;
+
+  return (
+    <div>
+      <div>{operationName}</div>
+      <div>{`${spanKind} - ${component}`}</div>
+    </div>
+  );
+}
